add fetchTxData and fetchTxLogs to blockchain/contracts
