+++ buyer-api/package.json
@@ -32,6 +32,7 @@
     "delay": "7000"
   },
   "dependencies": {
+    "abi-decoder": "^1.2.0",
     "apicache": "^1.2.3",
     "async-redis": "^1.1.4",
     "aws-sdk": "^2.188.0",
